+++ projects/ngx-nav-drawer/src/public_api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of ngx-nav-drawer
+ */
+
+export * from './ngx-nav-drawer.component';
+export * from './ngx-nav-drawer.module';