+++ src/app/project/multi-select/demo-multi-select-simple/demo-multi-select-simple.component.ts
@@ -8,10 +8,15 @@ import { Component } from '@angular/core';
 export class DemoMultiSelectSimpleComponent {
 
   selected = [];
-  options = ['South Africa', 'United States', 'United Kingdom'];
+  options = ['Afghanistan', 'Albania', 'Algeria', 'Andorra', 'Angola', 'Argentina', 'Armenia', 'Australia', 'Austria'];
+
+  get output() {
+    return JSON.stringify(this.selected);
+  }
 
   onChange(selection) {
     this.selected = selection;
   }
+
 
 }