+++ src/app/project/single-select/single-select.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-single-select',
+  templateUrl: './single-select.component.html',
+  styleUrls: ['./single-select.component.sass'],
+  preserveWhitespaces: true
+})
+export class SingleSelectComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}