+++ projects/ngx-clipboard/CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
-## 0.0.1
+## 8.0.2
+- docs update
+
+## 8.0.1
+- docs update
+
+## 8.0.0
 - initial release