+++ projects/ngx-code-highlight/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 8.0.3
+- docs update
+
 ## 8.0.1 - 8.0.2
 - Correct docs