+++ projects/ngx-multi-select/CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 
+## 8.0.4
+- Added Keyboard Navigation
+- Added Search for listed items
+
+## 8.0.3
+- improved styling styling
+
+## 8.0.2
+- docs update
+
 ## 8.0.1
-- Added theming
+- added custom styling
 
 ## 8.0.0
 - initial release