+++ projects/ngx-nav-drawer/CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog
 
-## 0.0.1
+## 8.0.0
 - initial release