+++ projects/ngx-quill-editor/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
 ## 8.0.1
+- docs update
+
+## 8.0.1
 - initial release