+++ projects/ngx-switch-input/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
-## 0.0.1
+## 8.0.1
+- style customization
+
+## 8.0.0
 - initial release